+++ lorawan/storage/service/async-wrapper-identity-service.cpp
@@ -57,8 +57,8 @@ void AsyncWrapperIdentityService::rm(
 }
 
 void AsyncWrapperIdentityService::list(
-    size_t offset,
-    size_t size,
+    uint32_t offset,
+    uint8_t size,
     std::function<void(
         int retCode,
         std::vector<NETWORKIDENTITY> &retval
