+++ lorawan/storage/service/async-wrapper-identity-service.h
@@ -47,8 +47,8 @@ public:
     );
 
     void list(
-        size_t offset,
-        size_t size,
+        uint32_t offset,
+        uint8_t size,
         std::function<void(
             int retCode,
             std::vector<NETWORKIDENTITY> &retval
