+++ lorawan/storage/service/identity-service-mem.cpp
@@ -4,7 +4,7 @@
 #include "lorawan/lorawan-error.h"
 #include "lorawan/lorawan-string.h"
 #include "lorawan/helper/file-helper.h"
-#include "lorawan/storage/serialization/identity-serialization.h"
+#include "lorawan/storage/serialization/identity-binary-serialization.h"
 
 #ifdef ESP_PLATFORM
 #include <iostream>
@@ -197,7 +197,7 @@ int ClientUDPIdentityService::cList(
 {
     IdentityListResponse r;
     r.response = list(r.identities, offset, size);
-    r.size = r.identities.size();
+    r.size = (uint8_t) r.identities.size();
     if (responseClient)
         responseClient->onIdentityList(nullptr, &r);
    return CODE_OK;
@@ -206,7 +206,7 @@ int ClientUDPIdentityService::cList(
 int ClientUDPIdentityService::cSize()
 {
     IdentityOperationResponse r;
-    r.size = size();
+    r.size = (uint8_t) size();
     if (responseClient)
         responseClient->onIdentityOperation(nullptr, &r);
     return CODE_OK;
